Move title to AssessmentForm
